fix: fix lint errors
