+++ .eslintrc
@@ -1,4 +1,9 @@
 {
     "extends": "airbnb",
-    "rules": { "react/jsx-filename-extension": [0] }
+    "rules": { "react/jsx-filename-extension": [0], "import/no-unresolved": [0], "class-methods-use-this": [0] },
+    "env": {
+        "browser": true,
+        "node": true,
+        "es6": true
+    }
 }
