+++ src/routes/index.js
@@ -1,6 +1,6 @@
 import { useRoutes } from 'react-router-dom';
 
-import { routes } from '@constants/routes';
+import routes from '@constants/routes';
 
 const RouterElement = () => useRoutes(routes);
