+++ src/utils/commands/AddCommand.js
@@ -1,0 +1,12 @@
+class AddCommand {
+  constructor(current, next) {
+    this.current = current;
+    this.next = next;
+  }
+
+  execute() {
+    return this.current + this.next;
+  }
+}
+
+export default AddCommand;
