+++ src/utils/commands/DivideCommand.js
@@ -1,0 +1,12 @@
+class DivideCommand {
+  constructor(current, next) {
+    this.current = current;
+    this.next = next;
+  }
+
+  execute() {
+    return this.current / this.next;
+  }
+}
+
+export default DivideCommand;
