+++ src/utils/commands/ModCommand.js
@@ -1,0 +1,12 @@
+class ModCommand {
+  constructor(current, next) {
+    this.current = current;
+    this.next = next;
+  }
+
+  execute() {
+    return this.current % this.next;
+  }
+}
+
+export default ModCommand;
