+++ src/utils/commands/MultiplyCommand.js
@@ -1,0 +1,12 @@
+class MultiplyCommand {
+  constructor(current, next) {
+    this.current = current;
+    this.next = next;
+  }
+
+  execute() {
+    return this.current * this.next;
+  }
+}
+
+export default MultiplyCommand;
