+++ src/utils/commands/SubtractCommand.js
@@ -1,0 +1,12 @@
+class SubtractCommand {
+  constructor(current, next) {
+    this.current = current;
+    this.next = next;
+  }
+
+  execute() {
+    return this.current - this.next;
+  }
+}
+
+export default SubtractCommand;
